Cleaned up kamvas_driver.py
- Rearranged a chunk of the code in kamvas_driver.py so that it will run
better as a standalone process. It is yet to be tested though so it is
likely to have a lot of bugs at this point
- Added some new config variables to config.yaml which specify the
mapping of raw usb action ids to actual actions
- Moved the function that used to be in kamvas_driver.py for parsing the
config. It still needs to be adjusted for parsing the new config
- Added options for running the driver service in the foreground for
debugging prurposes if the user needs to. The process will run in the
background by default